feat: Router API vercel handler
